Migrated withForm unit test to typescript
